[Form] moved data trimming logic of TrimListener into StringUtil
