+++ src/Symfony/Component/Form/Extension/Core/EventListener/TrimListener.php
@@ -14,6 +14,7 @@ namespace Symfony\Component\Form\Extension\Core\EventListener;
 use Symfony\Component\Form\FormEvents;
 use Symfony\Component\Form\FormEvent;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+use Symfony\Component\Form\Util\StringUtil;
 
 /**
  * Trims string data.
@@ -24,17 +25,7 @@ class TrimListener implements EventSubscriberInterface
 {
     public function preSubmit(FormEvent $event)
     {
-        $data = $event->getData();
-
-        if (!is_string($data)) {
-            return;
-        }
-
-        if (null !== $result = @preg_replace('/^[\pZ\p{Cc}]+|[\pZ\p{Cc}]+$/u', '', $data)) {
-            $event->setData($result);
-        } else {
-            $event->setData(trim($data));
-        }
+        $event->setData(StringUtil::trim($event->getData()));
     }
 
     /**
